% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ae_genes}
\alias{ae_genes}
\title{Adverse-event gene panel}
\format{
A character vector.
}
\source{
Curated manually (see \code{data-raw/make_internal_data.R}).
}
\usage{
ae_genes
}
\description{
A character vector of uppercase gene symbols flagged as adverse-event
associated. Used as the \code{AE_gene} source in
\code{\link{enrich_annotations}}.
}
\keyword{datasets}
